need to finish on Tuesday 3/12/19

+++ src/main/java/org/usfirst/frc/team1559/robot/Pixy.java
@@ -8,8 +8,8 @@ public class Pixy {
     private byte[] getFeatures = {(byte)0xae, (byte)0xc1, (byte)0x30, (byte)0x02, (byte)0x00, (byte)0x07, 0, 0};
     private byte[] lampon = {(byte)0xae, (byte)0xc1, (byte)0x16, (byte)0x02, (byte)0x01, (byte)0x01, 0, 0};
     private byte[] lampoff = {(byte)0xae, (byte)0xc1, (byte)0x16, (byte)0x02, (byte)0x00, (byte)0x00, 0, 0};
-    public float Ex;
-    public float Er;
+    public double Ex;
+    public double Er;
 	public Object v;
     public Pixy() {
         port = new SPI(SPI.Port.kOnboardCS1);
@@ -55,11 +55,11 @@ public class Pixy {
         }
         return v;
     }
-    public float getEx()
+    public double getEx()
         {
             return Ex;
         }
-    public float getEr()
+    public double getEr()
         {
             return Er;
         }
